a2amarketo/marketo_agent/requirements.txt
fastmcp>=2.0.0
google-adk>=1.0.0
marketorestpython>=0.5
fastapi>=0.95
uvicorn>=0.22.0
requests>=2.31.0
pydantic>=1.10
python-dotenv>=1.0.0